Refactor: 서버에 보내는 메시지만 return하도록 분리하여 websocketUtils.js 파일로 이동
