+++ FE/src/pages/DuelStatsPage.js
@@ -5,6 +5,7 @@ import DuelStatsData from '../components/DuelStatsData.js';
 import DuelBasicStats from '../components/DuelBasicStats.js';
 import DuelSpecialStats from '../components/DuelSpecialStats.js';
 import DuelGraphStats from '../components/DuelGraphStats.js';
+import { gameNextMatchMsg } from '../websocket/websocketUtils.js';
 
 const html = String.raw;
 
@@ -51,7 +52,7 @@ class DuelStatsPage {
 		const back = document.querySelector('.event-click-match');
 		back.addEventListener('click', () => {
 			if (this.data.Gamewebsocket) {
-				this.data.Gamewebsocket.sendGameNextMatch();
+				this.data.Gamewebsocket.send(gameNextMatchMsg());
 			} else changeUrl('match');
 		});
 	}
